adding volume year and volume id to the series name
Adding unconditionnally the volume start year and the volume id to the series name to be able to differentiate multiple series with the same name.
- with the volume id, we can manage the unicity
- with the volume year, we can order the series that have the same name.

When I have a little more time, I will add parameters so these additions can be added conditionnally and have this version be able to work like the old one.